+++ op-challenger/challenger/challenger.go
@@ -6,20 +6,24 @@ import (
 	"sync"
 	"time"
 
-	abi "github.com/ethereum/go-ethereum/accounts/abi"
-	bind "github.com/ethereum/go-ethereum/accounts/abi/bind"
-	common "github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
 	ethclient "github.com/ethereum/go-ethereum/ethclient"
-	log "github.com/ethereum/go-ethereum/log"
+	"github.com/ethereum/go-ethereum/log"
 
-	config "github.com/ethereum-optimism/optimism/op-challenger/config"
-	metrics "github.com/ethereum-optimism/optimism/op-challenger/metrics"
+	"github.com/ethereum-optimism/optimism/op-challenger/config"
+	"github.com/ethereum-optimism/optimism/op-challenger/metrics"
 
-	bindings "github.com/ethereum-optimism/optimism/op-bindings/bindings"
-	sources "github.com/ethereum-optimism/optimism/op-node/sources"
+	"github.com/ethereum-optimism/optimism/op-bindings/bindings"
+	"github.com/ethereum-optimism/optimism/op-node/eth"
 	opclient "github.com/ethereum-optimism/optimism/op-service/client"
-	txmgr "github.com/ethereum-optimism/optimism/op-service/txmgr"
+	"github.com/ethereum-optimism/optimism/op-service/txmgr"
 )
+
+type OutputAPI interface {
+	OutputAtBlock(ctx context.Context, blockNum uint64) (*eth.OutputResponse, error)
+}
 
 // Challenger contests invalid L2OutputOracle outputs
 type Challenger struct {
@@ -35,7 +39,7 @@ type Challenger struct {
 
 	l1Client *ethclient.Client
 
-	rollupClient *sources.RollupClient
+	rollupClient OutputAPI
 
 	// l2 Output Oracle contract
 	l2ooContract     *bindings.L2OutputOracleCaller
@@ -48,6 +52,33 @@ type Challenger struct {
 	dgfABI          *abi.ABI
 
 	networkTimeout time.Duration
+}
+
+// From returns the address of the account used to send transactions.
+func (c *Challenger) From() common.Address {
+	return c.txMgr.From()
+}
+
+// Client returns the client for the settlement layer.
+func (c *Challenger) Client() *ethclient.Client {
+	return c.l1Client
+}
+
+func (c *Challenger) NewOracleSubscription() (*Subscription, error) {
+	query, err := BuildOutputLogFilter(c.l2ooABI)
+	if err != nil {
+		return nil, err
+	}
+	return NewSubscription(query, c.Client(), c.log), nil
+}
+
+// NewFactorySubscription creates a new [Subscription] listening to the DisputeGameFactory contract.
+func (c *Challenger) NewFactorySubscription() (*Subscription, error) {
+	query, err := BuildDisputeGameLogFilter(c.dgfABI)
+	if err != nil {
+		return nil, err
+	}
+	return NewSubscription(query, c.Client(), c.log), nil
 }
 
 // NewChallenger creates a new Challenger