+++ op-node/p2p/peer_scorer_test.go
@@ -22,18 +22,14 @@ type PeerScorerTestSuite struct {
 	suite.Suite
 
 	mockStore    *p2pMocks.Peerstore
-	mockMetricer *p2pMocks.GossipMetricer
-	bandScorer   *p2p.BandScoreThresholds
+	mockMetricer *p2pMocks.ScoreMetrics
 	logger       log.Logger
 }
 
 // SetupTest sets up the test suite.
 func (testSuite *PeerScorerTestSuite) SetupTest() {
 	testSuite.mockStore = &p2pMocks.Peerstore{}
-	testSuite.mockMetricer = &p2pMocks.GossipMetricer{}
-	bandScorer, err := p2p.NewBandScorer("-40:graylist;0:friend;")
-	testSuite.NoError(err)
-	testSuite.bandScorer = bandScorer
+	testSuite.mockMetricer = &p2pMocks.ScoreMetrics{}
 	testSuite.logger = testlog.Logger(testSuite.T(), log.LvlError)
 }
 
@@ -48,7 +44,6 @@ func (testSuite *PeerScorerTestSuite) TestScorer_OnConnect() {
 		&rollup.Config{L2ChainID: big.NewInt(123)},
 		testSuite.mockStore,
 		testSuite.mockMetricer,
-		testSuite.bandScorer,
 		testSuite.logger,
 	)
 	scorer.OnConnect(peer.ID("alice"))
@@ -60,7 +55,6 @@ func (testSuite *PeerScorerTestSuite) TestScorer_OnDisconnect() {
 		&rollup.Config{L2ChainID: big.NewInt(123)},
 		testSuite.mockStore,
 		testSuite.mockMetricer,
-		testSuite.bandScorer,
 		testSuite.logger,
 	)
 	scorer.OnDisconnect(peer.ID("alice"))
@@ -72,19 +66,16 @@ func (testSuite *PeerScorerTestSuite) TestScorer_SnapshotHook() {
 		&rollup.Config{L2ChainID: big.NewInt(123)},
 		testSuite.mockStore,
 		testSuite.mockMetricer,
-		testSuite.bandScorer,
 		testSuite.logger,
 	)
 	inspectFn := scorer.SnapshotHook()
 
+	scores := store.PeerScores{Gossip: store.GossipScores{Total: 3}}
 	// Expect updating the peer store
-	testSuite.mockStore.On("SetScore", peer.ID("peer1"), &store.GossipScores{Total: float64(-100)}).Return(nil).Once()
+	testSuite.mockStore.On("SetScore", peer.ID("peer1"), &store.GossipScores{Total: float64(-100)}).Return(scores, nil).Once()
 
 	// The metricer should then be called with the peer score band map
-	testSuite.mockMetricer.On("SetPeerScores", map[string]float64{
-		"friend":   0,
-		"graylist": 1,
-	}).Return(nil).Once()
+	testSuite.mockMetricer.On("SetPeerScores", []store.PeerScores{scores}).Return(nil).Once()
 
 	// Apply the snapshot
 	snapshotMap := map[peer.ID]*pubsub.PeerScoreSnapshot{
@@ -95,48 +86,15 @@ func (testSuite *PeerScorerTestSuite) TestScorer_SnapshotHook() {
 	inspectFn(snapshotMap)
 
 	// Expect updating the peer store
-	testSuite.mockStore.On("SetScore", peer.ID("peer1"), &store.GossipScores{Total: 0}).Return(nil).Once()
+	testSuite.mockStore.On("SetScore", peer.ID("peer1"), &store.GossipScores{Total: 0}).Return(scores, nil).Once()
 
 	// The metricer should then be called with the peer score band map
-	testSuite.mockMetricer.On("SetPeerScores", map[string]float64{
-		"friend":   1,
-		"graylist": 0,
-	}).Return(nil).Once()
+	testSuite.mockMetricer.On("SetPeerScores", []store.PeerScores{scores}).Return(nil).Once()
 
 	// Apply the snapshot
 	snapshotMap = map[peer.ID]*pubsub.PeerScoreSnapshot{
 		peer.ID("peer1"): {
 			Score: 0,
-		},
-	}
-	inspectFn(snapshotMap)
-}
-
-// TestScorer_SnapshotHookBlocksPeer tests running the snapshot hook on the peer scorer with a peer score below the threshold.
-// This implies that the peer should be blocked.
-func (testSuite *PeerScorerTestSuite) TestScorer_SnapshotHookBlocksPeer() {
-	scorer := p2p.NewScorer(
-		&rollup.Config{L2ChainID: big.NewInt(123)},
-		testSuite.mockStore,
-		testSuite.mockMetricer,
-		testSuite.bandScorer,
-		testSuite.logger,
-	)
-	inspectFn := scorer.SnapshotHook()
-
-	// Expect updating the peer store
-	testSuite.mockStore.On("SetScore", peer.ID("peer1"), &store.GossipScores{Total: float64(-101)}).Return(nil).Once()
-
-	// The metricer should then be called with the peer score band map
-	testSuite.mockMetricer.On("SetPeerScores", map[string]float64{
-		"friend":   0,
-		"graylist": 1,
-	}).Return(nil)
-
-	// Apply the snapshot
-	snapshotMap := map[peer.ID]*pubsub.PeerScoreSnapshot{
-		peer.ID("peer1"): {
-			Score: -101,
 		},
 	}
 	inspectFn(snapshotMap)